Create starsign.py
I've been trying to make this piece of code work for a while now, but I've encountered numerous errors I can't fix. I use Visual Studio Code and clearly the code is written in python. 
This is the message I get when I try to run the code:
[Running] python -u "c:\Users\Candidate\Documents\Hello World\starsigns.py"
'python' is not recognized as an internal or external command,
operable program or batch file.

[Done] exited with code=1 in 0.075 seconds